requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
python-dotenv
alembic
pydantic
psycopg2-binary